pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "superset-mcp"
version = "0.1.0"
description = "MCP server for Apache Superset"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
superset-mcp = "superset_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["superset_mcp"]
